requirements.txt
apscheduler==3.10.4
httpx==0.25.2
openai==1.13.3
python-dotenv==1.0.1
python-telegram-bot==20.7
rich==13.7.1
typer==0.11.0
uvloop==0.19.0; sys_platform != "win32"
pytest==9.0.1
